Move include files outside web root
