+++ admin/reindex.php
@@ -12,7 +12,8 @@
 */
 require_once '../../../lib-common.php';
 require_once '../../auth.inc.php';
-require_once 'admin.inc.php';
+require_once $_CONF['path'].'plugins/searcher/include/admin.inc.php';
+require_once $_CONF['path'].'plugins/searcher/include/reindex.ajax.php';
 
 USES_lib_admin();
 
@@ -22,7 +23,7 @@ function SRCHER_reindex()
 
     $retval = '';
 
-    $T = new Template(SRCH_PI_PATH . '/templates');
+    $T = new \Template(SRCH_PI_PATH . '/templates');
     $T->set_file('page','reindex.thtml');
 
     $retval .= SRCH_adminMenu('reindex');
@@ -66,125 +67,7 @@ function SRCHER_reindex()
 }
 
 
-function SRCH_getContentTypesAjax()
-{
-    global $_PLUGINS;
 
-    if ( !COM_isAjax()) die();
-
-    $contentTypes = array();
-    $retval = array();
-
-    $contentTypes[] = 'article';
-
-    foreach ($_PLUGINS as $pi_name) {
-        if (function_exists('plugin_getiteminfo_' . $pi_name)) {
-            $contentTypes[] = $pi_name;
-        }
-    }
-
-    $retval['errorCode'] = 0;
-    $retval['contenttypes'] = $contentTypes;
-
-    $retval['statusMessage'] = 'Initialization Successful';
-
-    $return["js"] = json_encode($retval);
-
-    echo json_encode($return);
-    exit;
-}
-
-
-function SRCH_getContentListAjax()
-{
-    global $_PLUGINS;
-
-    if ( !COM_isAjax()) die();
-
-    USES_searcher_class_indexer();
-
-    if ( !isset($_POST['type'])) die();
-
-    $type = COM_applyFilter($_POST['type']);
-
-    $contentList = array();
-    $retval = array();
-
-    $rc = PLG_getItemInfo($type,'*','id,search_index');
-    foreach ( $rc AS $id ) {
-        $contentList[] = $id;
-    }
-
-    \Searcher\Indexer::Removeall($type);
-
-    $retval['errorCode'] = 0;
-    $retval['contentlist'] = $contentList;
-    $retval['statusMessage'] = 'Content List Successful';
-
-    $return["js"] = json_encode($retval);
-
-    echo json_encode($return);
-    exit;
-}
-
-function SRCH_indexContentItemAjax()
-{
-    global $_PLUGINS;
-
-    if ( !COM_isAjax()) die();
-
-    if ( !isset($_POST['type'])) die();
-    if ( !isset($_POST['id'])) die();
-
-    USES_searcher_class_indexer();
-
-    $type = COM_applyFilter($_POST['type']);
-    $id   = COM_applyFilter($_POST['id']);
-
-    $contentList = array();
-    $retval = array();
-
-    $contentInfo = PLG_getItemInfo($type,$id,'id,date,title,searchidx,author,hits,perms,search_index,reindex');
-
-    if ( is_array($contentInfo) && count($contentInfo) > 0 ) {
-        $props = array(
-            'item_id' => $id,
-            'type'  => $type,
-            'title' => $contentInfo['title'],
-            'content' => $contentInfo['searchidx'],
-            'date' => $contentInfo['date'],
-            'perms' => array(
-                'owner_id' => $contentInfo['perms']['owner_id'],
-                'group_id' => $contentInfo['perms']['group_id'],
-                'perm_owner' => $contentInfo['perms']['perm_owner'],
-                'perm_group' => $contentInfo['perms']['perm_group'],
-                'perm_members' => $contentInfo['perms']['perm_members'],
-                'perm_anon' => $contentInfo['perms']['perm_anon'],
-            ),
-        );
-
-        \Searcher\Indexer::IndexDoc($props);
-
-        if (function_exists('plugin_commentsupport_'.$type ) ) {
-            $func = 'plugin_commentsupport_'.$type;
-            $rc = $func();
-            if ( $rc == true ) {
-                plugin_IndexAll_comments($type, $id, $props['perms']);
-            }
-        }
-
-        $retval['errorCode'] = 0;
-        $retval['statusMessage'] = 'Content Item Index Successful';
-    } else {
-        $retval['errorCode'] = -1;
-        $retval['statusMessage'] = 'Error indexing content';
-    }
-
-    $return["js"] = json_encode($retval);
-
-    echo json_encode($return);
-    exit;
-}
 
 // main driver
 $action = '';
